second commit

+++ firstprog.java
@@ -9,7 +9,7 @@ public class firstprogram
 		int c = max(a,b);
 	}
 
-	int max(int a, int b)
+	float max(int a, int b)
 	{
 		if(a>b)
 		{
